CartCounter StateChange fix and cleanup
Counter was resetting to 0 every time cart items were updated, due to a re-fetch from api.  Connected it to redux to get global state and preserve count between cart updates.  (So no longer flashes "Cart(0)" between loads).

Leaving in the GetCartItems fetch to api on Mount - for now - as I'm not sure if it's better to fetch only when the cart counter is needed (which is every page to be fair) and risk multiple instances of the component calling it repeatedly, or put it somewhere global like App and at least know it's only called once.  TODO.